pyproject.toml
[tool.poetry]
name = "checkbox-test"
version = "0.1.0"
description = ""
authors = ["Ihor Moroz <igorm.200323@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["all"], version = "0.115.10"}
sqlalchemy = "2.0.38"
asyncpg = "0.30.0"
alembic = "1.14.1"
python-jose = {extras = ["cryptography"], version = "3.4.0"}
redis = "5.2.1"
bcrypt = "4.3.0"
fastapi-limiter = "0.1.6"
greenlet = "3.1.1"
passlib = {extras = ["bcrypt"], version = "1.7.4"}
qrcode = "8.0"
pillow = "11.1.0"

[tool.poetry.group.dev.dependencies]
sphinx = "8.2.3"
pytest-asyncio = "0.25.3"
fakeredis = "2.27.0"


[tool.poetry.group.test.dependencies]
pytest = "8.3.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
